SPOI-6446: Merge PojoEnrichment and TupleEnrichment
Reverted All the operators for since tag
Deleted original POJoEnrichmentOperator
Moved TupleEnrichmentOperator to POJOEnrichmentOperator
Made POJOEnrichmentOperator as Envolving.
